% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status_datasets.R
\name{check_status_datasets}
\alias{check_status_datasets}
\title{Check the status of local Glottography datasets}
\usage{
check_status_datasets(datasets = "all", online = TRUE)
}
\arguments{
\item{datasets}{Character vector. Names of the dataset(s) to check, or \code{"all"} to check all datasets. Defaults to \code{"all"}.}

\item{online}{Logical. Should the function attempt to check online for the latest versions? Defaults to TRUE.}
}
\value{
Invisible list of three data.frames:
\itemize{
\item \code{not_installed}: datasets available online but not downloaded locally
\item \code{outdated}: datasets installed but not up-to-date
\item \code{up_to_date}: datasets that are installed and up-to-date
}
}
\description{
Checks if specified Glottography datasets are present locally and whether they are up-to-date.
}
