% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_registry.R
\name{.get_registry}
\alias{.get_registry}
\title{Retrieves the (synced) local registry file}
\usage{
.get_registry(sync = TRUE)
}
\arguments{
\item{sync}{(logical) ensures that local registry is synced with Zenodo}
}
\value{
(synced) local registry as data.frame
}
\description{
Retrieves the (synced) local registry file
}
