% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download one or several Glottography dataset(s)}
\usage{
download_dataset(name)
}
\arguments{
\item{name}{the name(s) of the dataset(s) to download}

\item{check_for_updates}{verify that the local registry is up to date}
}
\description{
Takes the name of a data set as input, checks if it is in the local registry.
If yes, downloads the corresponding Glottography dataset from Zenodo. Can also verify if the local registry is up to date
}
