% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List available Glottography datasets}
\usage{
list_datasets()
}
\value{
A data.frame containing dataset metadata.
}
\description{
Reads the local registry of available Glottography datasets and
returns a data.frame with metadata such as title, version, and DOI.
}
