% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List available Glottography datasets}
\usage{
list_datasets(online = TRUE)
}
\arguments{
\item{online}{Logical. Should the function attempt to sync the registry with
the online version? Defaults to \code{TRUE}.}
}
\value{
A \code{data.frame} containing metadata and status information for all
Glottography datasets with the following columns:
\itemize{
\item \code{name}: Dataset name. Serves as a unique identifier.
\item \code{installed}: Logical; whether the dataset is installed locally.
\item \code{created}: Creation date of the dataset on Zenodo.
\item \code{version}: Latest available version online.
\item \code{version_local}: Version currently installed locally, or \code{NA} if none.
\item \code{modified}: Timestamp of the most recent online update.
\item \code{modified_local}: Timestamp of the local version, or \code{NA} if none.
}
}
\description{
Lists all available Glottography datasets.
The function reads the package’s registry file, which contains metadata for all
datasets along information on local installation status.
}
\details{
If \code{online = TRUE} and an internet connection is available, the function
attempts to sync the registry with the Glottography community on Zenodo by
updating the local registry. Otherwise, it runs in offline mode and uses
the local, unsynced registry.
}
